+++ packages/cms-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aphexcms/cms-core
 
+## 5.1.0
+
+### Minor Changes
+
+- added a bunch of fixes
+
+## 5.0.6
+
+### Patch Changes
+
+- Added vite plugin for HMR - upgradable
+
 ## 5.0.5
 
 ### Patch Changes